FinalScheduler/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pdfplumber
PyPDF2
pypdf
pymupdf
pandas
openpyxl
numpy
requests
reportlab
pymongo
readline